React Router 1. Fixing react router not display path 2. install react router 3. add Bottom Navigation Component
