+++ src/App.js
@@ -7,7 +7,7 @@ function App(){
     <React.Fragment>
         <Header/>
         <Main/>
-        {/*<Footer/>*/}
+        <Footer/>
     </React.Fragment>
   )
 }
